refactor from pr comments
