+++ MovieAppPetProject/MovieAppPetProject/Base.lproj/Main.storyboard
@@ -45,7 +45,7 @@
                 </navigationController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="Cb4-VH-9lC" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="-995" y="574"/>
+            <point key="canvasLocation" x="-1215" y="574"/>
         </scene>
         <!--HomeScreen-->
         <scene sceneID="N6j-Hf-gJn">
@@ -55,7 +55,7 @@
                 </viewControllerPlaceholder>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="pPp-kA-edD" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="-996" y="975"/>
+            <point key="canvasLocation" x="-1217" y="997"/>
         </scene>
         <!--Search-->
         <scene sceneID="fSu-vq-gjG">
@@ -72,13 +72,13 @@
                 </navigationController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="Q6o-Tu-LcB" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="236" y="574"/>
+            <point key="canvasLocation" x="-384" y="574"/>
         </scene>
-        <!--Item-->
+        <!--Watchlist-->
         <scene sceneID="AL8-WJ-8GN">
             <objects>
                 <navigationController id="vGD-DL-K23" sceneMemberID="viewController">
-                    <tabBarItem key="tabBarItem" title="Item" id="eqH-pI-hVC"/>
+                    <tabBarItem key="tabBarItem" title="Watchlist" image="popcorn.circle.fill" catalog="system" id="eqH-pI-hVC"/>
                     <navigationBar key="navigationBar" contentMode="scaleToFill" id="WuT-gn-w2e">
                         <rect key="frame" x="0.0" y="59" width="393" height="44"/>
                         <autoresizingMask key="autoresizingMask"/>
@@ -89,7 +89,7 @@
                 </navigationController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="RUB-Sh-pdn" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="1222" y="574"/>
+            <point key="canvasLocation" x="411" y="574"/>
         </scene>
         <!--WatchlistScreen-->
         <scene sceneID="bvS-Ai-0J6">
@@ -99,7 +99,7 @@
                 </viewControllerPlaceholder>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="2M1-Kh-8g0" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="1220" y="985"/>
+            <point key="canvasLocation" x="410" y="997"/>
         </scene>
         <!--SearchMovieScreen-->
         <scene sceneID="iNa-YH-Wcy">
@@ -109,12 +109,13 @@
                 </viewControllerPlaceholder>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="YPN-97-yxO" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="236" y="975"/>
+            <point key="canvasLocation" x="-386" y="997"/>
         </scene>
     </scenes>
     <resources>
         <image name="house.circle.fill" catalog="system" width="128" height="123"/>
         <image name="magnifyingglass.circle.fill" catalog="system" width="128" height="123"/>
+        <image name="popcorn.circle.fill" catalog="system" width="128" height="123"/>
         <systemColor name="systemIndigoColor">
             <color red="0.34509803919999998" green="0.33725490200000002" blue="0.83921568629999999" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
         </systemColor>
