+++ MovieAppPetProject/MovieAppPetProject/CoreDataFiles/CoreDataManager.swift
@@ -5,7 +5,6 @@
 //  Created by Sebastian Jacobs on 2024/05/14.
 //
 
-import Foundation
 import UIKit
 
 // MARK: - Enum
@@ -18,7 +17,7 @@ class CoreDataManager {
     let context = (UIApplication.shared.delegate as? AppDelegate)?.persistentContainer.viewContext
 
     // MARK: - Functions
-    func getAllWatchlistItems() -> [WatchList] {
+    func fetchAllWatchlistItems() -> [WatchList] {
         do {
             guard let context = context else { throw CoreDataError.noContext }
             return try context.fetch(WatchList.fetchRequest())
