+++ MovieAppPetProject/MovieAppPetProject/CoreDataFiles/Model.xcdatamodeld/Model.xcdatamodel/contents
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="22522" systemVersion="23C71" minimumToolsVersion="Automatic" sourceLanguage="Swift" usedWithSwiftData="YES" userDefinedModelVersionIdentifier="">
-    <entity name="WatchList" representedClassName="WatchList" syncable="YES">
+    <entity name="WatchList" representedClassName="WatchList" syncable="YES" codeGenerationType="class">
         <attribute name="originalTitle" optional="YES" attributeType="String"/>
     </entity>
 </model>
