--- MovieAppPetProject/MovieAppPetProject/CoreDataFiles/WatchList+CoreDataClass.swift
@@ -1,15 +1,0 @@
-//
-//  WatchList+CoreDataClass.swift
-//  MovieAppPetProject
-//
-//  Created by Sebastian Jacobs on 2024/05/14.
-//
-//
-
-import Foundation
-import CoreData
-
-@objc(WatchList)
-public class WatchList: NSManagedObject {
-
-}
