+++ MovieAppPetProject/MovieAppPetProject/GetMovieDetailsByID/GetMovieDetails-ViewController.swift
@@ -20,12 +20,12 @@ class MovieDetailsViewController: UIViewController {
     @IBOutlet private weak var status: UILabel!
 
     // MARK: - IBAction
-    @IBAction func addToWatchlistTapped(_ sender: UIButton) {
+    @IBAction private func addToWatchlistTapped(_ sender: UIButton) {
         movieDetailsViewModel.addToWatchlist()
     }
 
     // MARK: - Variables
-    private lazy var movieDetailsViewModel = MovieDetailsViewModel(movieDetailsRepository: MovieDetailsRepository())
+    private lazy var movieDetailsViewModel = MovieDetailsViewModel(movieDetailsRepository: MovieDetailsRepository(coreDataManager: CoreDataManager()))
 
     // MARK: - Functions
     override func viewDidLoad() {
